Pasa de nivel, hay que mejorarlo cuando hay dorados

+++ Arkanoid/src/app/Movible.java
@@ -2,6 +2,6 @@ package app;
 
 public interface Movible{
     public void mover() ;  //mover
-    public double velocidad() ;//velocidad
-    public double aceleracion() ;//aceleracion
+    public int velocidad() ;//velocidad
+    public int aceleracion() ;//aceleracion
 }
